+++ lib/Hasql/Interpolate/Internal/Json.hs
@@ -4,6 +4,8 @@
 module Hasql.Interpolate.Internal.Json
   ( Json (..),
     Jsonb (..),
+    JsonBytes (..),
+    JsonbBytes (..),
     AsJson (..),
     AsJsonb (..),
   )
@@ -12,6 +14,7 @@ where
 import Data.Aeson
 import qualified Data.Aeson as Aeson
 import Data.Bifunctor (first)
+import Data.ByteString (ByteString)
 import qualified Data.ByteString.Lazy as BL
 import Data.Coerce
 import Data.Functor.Contravariant
@@ -23,23 +26,33 @@ import Hasql.Interpolate.Internal.Encoder
 
 -- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
 -- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
--- between a postgres json type and an Aeson 'Value'
+-- between a postgres @jsonb@ and an Aeson 'Value'
 newtype Jsonb = Jsonb Value
 
 -- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
 -- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
--- between a postgres json type and an Aeson 'Value'
+-- between a postgres @json@ and an Aeson 'Value'
 newtype Json = Json Value
 
 -- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
 -- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
--- between a postgres json type and anything that is an instance of
+-- between a postgres @jsonb@ and a 'ByteString'
+newtype JsonbBytes = JsonbBytes ByteString
+
+-- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
+-- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
+-- between a postgres @json@ and a 'ByteString'
+newtype JsonBytes = JsonBytes ByteString
+
+-- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
+-- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
+-- between a postgres @json@ and anything that is an instance of
 -- 'FromJSON' / 'ToJSON'
 newtype AsJson a = AsJson a
 
 -- | Newtype for 'Hasql.Interpolate.Decoder.DecodeValue' /
 -- 'Hasql.Interpolate.Encoder.EncodeValue' instances that converts
--- between a postgres jsonb type and anything that is an instance of
+-- between a postgres @jsonb@ and anything that is an instance of
 -- 'FromJSON' / 'ToJSON'
 newtype AsJsonb a = AsJsonb a
 
@@ -50,6 +63,14 @@ instance DecodeValue Jsonb where
 -- | Parse a postgres @json@ using 'D.json'
 instance DecodeValue Json where
   decodeValue = coerce D.json
+
+-- | Parse a postgres @jsonb@ using 'D.jsonbBytes'
+instance DecodeValue JsonbBytes where
+  decodeValue = coerce (D.jsonbBytes Right)
+
+-- | Parse a postgres @json@ using 'D.jsonBytes'
+instance DecodeValue JsonBytes where
+  decodeValue = coerce (D.jsonBytes Right)
 
 -- | Parse a postgres @json@ to anything that is an instance of
 -- 'Aeson.FromJSON'
@@ -68,6 +89,14 @@ instance EncodeValue Json where
 -- | Encode an Aeson 'Aeson.Value' to a postgres @jsonb@ using 'E.jsonb'
 instance EncodeValue Jsonb where
   encodeValue = coerce E.jsonb
+
+-- | Encode a 'ByteString' to a postgres @json@ using 'E.jsonBytes'
+instance EncodeValue JsonBytes where
+  encodeValue = coerce E.jsonbBytes
+
+-- | Encode a 'ByteString' to a postgres @jsonb@ using 'E.jsonbBytes'
+instance EncodeValue JsonbBytes where
+  encodeValue = coerce E.jsonbBytes
 
 -- | Encode anything that is an instance of 'Aeson.ToJSON' to a postgres @json@
 instance Aeson.ToJSON a => EncodeValue (AsJson a) where